+++ src/common/constants/constants.ts
@@ -3,7 +3,7 @@ export const PAGES_PATHS = {
   SIGN_UP: '/signup',
   HOME: '/',
   RESET_PASS: '/reset-password',
-  RESULTS: '/results',
+  MANAGE: '/manage',
 }
 
 export const BUTTON_LABELS = {
