--- src/features/manage/components/manage-description/manage-description.tsx
@@ -1,3 +1,0 @@
-export const ManageDescription = () => {
-  return <p>manage-description</p>
-}
