--- src/features/manage/components/manage-details/manage-details.tsx
@@ -1,3 +1,0 @@
-export const ManageDetails = () => {
-  return <p>manage-details</p>
-}
