+++ src/features/manage/components/manage-photos/manage-photos.module.scss
@@ -1,0 +1,79 @@
+@import 'styles/convert.scss';
+
+.managePhotos {
+  display: grid;
+  width: 100%;
+  grid-template-columns: 2fr 1fr 1fr;
+  gap: rem(30px);
+  grid-template-areas:
+    'grid1 grid2 grid3'
+    'grid1 grid4 grid5';
+  position: relative;
+}
+
+.managePhotosGrid1 {
+  grid-area: grid1;
+}
+
+.managePhotosGrid2 {
+  grid-area: grid2;
+  width: 100%;
+}
+
+.managePhotosGrid3 {
+  grid-area: grid3;
+  width: 100%;
+  border-radius: 0px rem(12px) 0px 0px;
+}
+
+.managePhotosGrid4 {
+  grid-area: grid4;
+  width: 100%;
+}
+
+.managePhotosGrid5 {
+  grid-area: grid5;
+  width: 100%;
+}
+
+.managePhotosGrid1Style {
+  grid-area: grid1;
+  width: 100%;
+  height: 100%;
+  border-radius: rem(12px) 0px 0px rem(12px);
+  object-fit: cover;
+}
+
+.managePhotosGrid3Style {
+  width: 100%;
+  border-radius: 0px rem(12px) 0px 0px;
+}
+
+.managePhotosGrid5Style {
+  width: 100%;
+  border-radius: 0px 0px rem(12px) 0px;
+}
+
+.manageAllPhotoButton {
+  width: rem(160px);
+  height: rem(36px);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: rem(12px);
+  background-color: white;
+  border: rem(1px) solid var(--primary-500);
+  border-radius: rem(8px);
+  position: absolute;
+  bottom: rem(12px);
+  right: rem(12px);
+  cursor: pointer;
+}
+
+.manageAllPhoto {
+  font-style: normal;
+  font-weight: 600;
+  font-size: rem(14px);
+  line-height: rem(22px);
+  color: var(--primary-500);
+}
