--- src/features/results/components/results-photos/results-photos.module.scss
@@ -1,79 +1,0 @@
-@import 'styles/convert.scss';
-
-.proprety {
-  display: grid;
-  width: 100%;
-  grid-template-columns: 2fr 1fr 1fr;
-  gap: rem(30px);
-  grid-template-areas:
-    'grid1 grid2 grid3'
-    'grid1 grid4 grid5';
-  position: relative;
-}
-
-.grid1 {
-  grid-area: grid1;
-}
-
-.grid2 {
-  grid-area: grid2;
-  width: 100%;
-}
-
-.grid3 {
-  grid-area: grid3;
-  width: 100%;
-  border-radius: 0px rem(12px) 0px 0px;
-}
-
-.grid4 {
-  grid-area: grid4;
-  width: 100%;
-}
-
-.grid5 {
-  grid-area: grid5;
-  width: 100%;
-}
-
-.grid1Style {
-  grid-area: grid1;
-  width: 100%;
-  height: 100%;
-  border-radius: rem(12px) 0px 0px rem(12px);
-  object-fit: cover;
-}
-
-.grid3Style {
-  width: 100%;
-  border-radius: 0px rem(12px) 0px 0px;
-}
-
-.grid5Style {
-  width: 100%;
-  border-radius: 0px 0px rem(12px) 0px;
-}
-
-.allPhotoButton {
-  width: rem(160px);
-  height: rem(36px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: rem(12px);
-  background-color: white;
-  border: rem(1px) solid var(--primary-500);
-  border-radius: rem(8px);
-  position: absolute;
-  bottom: rem(12px);
-  right: rem(12px);
-  cursor: pointer;
-}
-
-.allPhoto {
-  font-style: normal;
-  font-weight: 600;
-  font-size: rem(14px);
-  line-height: rem(22px);
-  color: var(--primary-500);
-}
