--- src/features/results/constants/results-constants.ts
@@ -1,23 +1,0 @@
-export const PROPRETY_ALT_IMG = {
-  IMAGE_HOME: 'Home Photo',
-  SHOW: 'Show all photo',
-  AVATAR: 'User Photo',
-  SHARE: 'Share Icon',
-  FAVORITE: 'Favorite Icon',
-}
-
-export const PROPRETY_BUTTON = {
-  ALL_IMG: 'Show all photos',
-  PURCHASE: 'Purchase',
-  DESCRIPTION: 'Description',
-  SHARE: 'Share',
-  SEND: 'Send',
-}
-
-export const PROPRETY_LABELS = {
-  MESSAGES_LABELS: 'Message the seller',
-}
-
-export const PROPRETY_PLACEHOLDER = {
-  TEXTAREA: 'Messages..',
-}
